Extract ContextBuilder event-translation seam from GovernancePipeline
Move the four event-to-context translators (context_from_session_event,
enrich_event, and the shell helpers _classify_shell_for_assess /
_build_command_analysis) into a ContextBuilder collaborator holding only the
classification engine and project root. It is a pure adapter: SessionEvent /
ToolCallEvent in, EnrichmentContext out, no session state, no store.

The pipeline constructs self._context and keeps context_from_session_event /
enrich_event as thin self-dispatched forwards, so external callers and the
test_bridge fail-closed monkeypatch continue to work. Literal is no longer
used in pipeline and was dropped from its typing import. No behavior change:
1884 passed / 4 skipped.
